refactor: app creation handler
- decouple template processing
- add new templates
