+++ bin/cli.js
@@ -1,12 +1,10 @@
 #!/usr/bin/env node
 
 const yargs = require('yargs')
-const { newApp } = require('../lib/cli/handlers')
+const { newApp } = require('../lib/cli/commands')
 
-const argv = yargs
+yargs
     .command('new [app]', 'Create a new application', {}, newApp)
     .help('h')
     .alias('h', 'help')
     .argv
-
-// TODO handle arguments
