+++ lib/cli/commands.js
@@ -1,10 +1,10 @@
 const fs = require('fs')
 const child = require('child_process')
-const handlebars = require('handlebars')
 const logger = require('./logger')
+const templates = require('./templates')
 const prompt = require('prompt-sync')({ sigint: true })
 
-function createAppFolders(app) {
+function createDefaultFolders(app) {
     const rootPath = `${process.cwd()}/${app}`
 
     try {
@@ -25,6 +25,13 @@ function createAppFolders(app) {
     return rootPath
 }
 
+function copyDefaultFiles(rootPath, appName) {
+    fs.copyFileSync(
+        `${templates.path}/tsconfig.json`,
+        `${rootPath}/tsconfig.json`
+    )
+}
+
 function inquireAppContext(appName) {
     logger.debug('Inquirying project information')
 
@@ -38,15 +45,13 @@ function inquireAppContext(appName) {
 function newApp(argv) {
     logger.info('Setting up your awesome project...')
 
-    const appRoot = createAppFolders(argv.app)
+    const appRoot = createDefaultFolders(argv.app)
+    copyDefaultFiles(appRoot, argv.app)
 
-    const packageTemplate = `${__dirname}/templates/package.json.hbs`
-    const content = fs.readFileSync(packageTemplate, 'utf8')
     const context = inquireAppContext(argv.app)
+    const packageJson = templates.parse('package.json.hbs', context)
 
     const appPackage = `${appRoot}/package.json`
-    const packageJson = handlebars.compile(content)(context)
-
     fs.writeFileSync(appPackage, packageJson)
 
     logger.info('Installing dependencies...')
