+++ lib/cli/templates/README.md.hbs
@@ -1,0 +1,9 @@
+# {{ appName }}
+
+You should include useful information in here!
+
+For exaple:
+
+- How to install it
+- How to run it
+- Contribution guidelines for developers
