+++ lib/cli/templates/package.json.hbs
@@ -6,6 +6,7 @@
   "author": "{{ appAuthor }}",
   "license": "{{ appLicence }}",
   "dependencies": {
+    "typedi": "0.8.0",
     "typescript": "3.8.3"
   },
   "engines": {
