+++ lib/cli/templates/tsconfig.json
@@ -1,0 +1,25 @@
+{
+    "compilerOptions": {
+        "outDir": "build/",
+        "baseUrl": ".",
+        "declaration": true,
+        "typeRoots": ["node_modules/@types"],
+        "allowJs": false,
+        "allowSyntheticDefaultImports": true,
+        "esModuleInterop": true,
+        "lib": ["es7"],
+        "moduleResolution": "node",
+        "strict": true,
+        "module": "commonjs",
+        "target": "es6",
+        "strictNullChecks": true,
+        "noImplicitThis": true,
+        "noImplicitReturns": true,
+        "noUnusedLocals": true,
+        "noUnusedParameters": true,
+        "skipLibCheck": true,
+        "noImplicitAny": true
+    },
+    "include": ["./src/*"],
+    "exclude": ["node_modules"]
+}
